requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
dlib==20.0.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Faker==37.12.0
idna==3.11
numpy==2.3.4
openpyxl==3.1.5
pillow==12.0.0
PyMySQL==1.1.2
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
